Split menu bar and toolbars in serarate files and subclasses. Moved grab data, load and save methods to corresponding sections and files
